+++ firebase-vscode/src/data-connect/language-client.ts
@@ -7,12 +7,15 @@ import {
   LanguageClient,
 } from "vscode-languageclient/node";
 import * as path from "node:path";
-import { Signal } from "@preact/signals-core";
+import { ResolvedDataConnectConfigs } from "./config";
 
-export function setupLanguageClient(context) {
+export function setupLanguageClient(
+  context,
+  configs: ResolvedDataConnectConfigs,
+) {
   // activate language client/serer
   const outputChannel: vscode.OutputChannel = vscode.window.createOutputChannel(
-    "Firebase GraphQL Language Server"
+    "Firebase GraphQL Language Server",
   );
 
   const serverPath = path.join("dist", "server.js");
@@ -47,12 +50,12 @@ export function setupLanguageClient(context) {
           // also, it makes sense that it should only re-load on file save, but we need to document that.
           // TODO: perhaps we can intercept change events, and remind the user
           // to save for the changes to take effect
-          true
+          true,
         ),
         // TODO: load ignore file
         // These ignore node_modules and .git by default
         vscode.workspace.createFileSystemWatcher(
-          "**/{*.graphql,*.graphqls,*.gql,*.js,*.mjs,*.cjs,*.esm,*.es,*.es6,*.jsx,*.ts,*.tsx,*.vue,*.svelte,*.cts,*.mts}"
+          "**/{*.graphql,*.graphqls,*.gql,*.js,*.mjs,*.cjs,*.esm,*.es,*.es6,*.jsx,*.ts,*.tsx,*.vue,*.svelte,*.cts,*.mts}",
         ),
       ],
     },
@@ -75,38 +78,70 @@ export function setupLanguageClient(context) {
     "graphQLlanguageServer",
     "GraphQL Language Server",
     serverOptions,
-    clientOptions
+    clientOptions,
   );
 
   // register commands
   const commandShowOutputChannel = vscode.commands.registerCommand(
     "fdc-graphql.showOutputChannel",
-    () => outputChannel.show()
+    () => outputChannel.show(),
   );
 
   context.subscriptions.push(commandShowOutputChannel);
 
+  const generateYamlFile = async () => {
+    const basePath = vscode.workspace.rootPath;
+    const filePath = ".firebase/.graphqlrc";
+    const fileUri = vscode.Uri.file(`${basePath}/${filePath}`);
+    const folderPath = ".firebase";
+    const folderUri = vscode.Uri.file(`${basePath}/${folderPath}`);
+
+    // TODO: Expand to multiple services
+    const config = configs.values[0];
+    const generatedPath = ".dataconnect";
+    const schemaPaths = [
+      `../${config.relativeSchemaPath}/**/*.gql`,
+      `../${config.relativePath}/${generatedPath}/**/*.gql`,
+    ];
+    const documentPaths = config.relativeConnectorPaths.map(
+      (connectorPath) => `../${connectorPath}/**/*.gql`,
+    );
+
+    const yamlJson = JSON.stringify({
+      schema: schemaPaths,
+      document: documentPaths,
+    });
+    // create folder if needed
+    if (!vscode.workspace.getWorkspaceFolder(folderUri)) {
+      vscode.workspace.fs.createDirectory(folderUri);
+    }
+    vscode.workspace.fs.writeFile(fileUri, Buffer.from(yamlJson));
+  };
+
   vscode.commands.registerCommand("fdc-graphql.restart", async () => {
     outputChannel.appendLine("Stopping Firebase GraphQL Language Server");
     await client.stop();
-
+    await generateYamlFile();
     outputChannel.appendLine("Restarting Firebase GraphQL Language Server");
     await client.start();
     outputChannel.appendLine("Firebase GraphQL Language Server restarted");
   });
 
-  const restartGraphqlLSP = () => {
-    vscode.commands.executeCommand("fdc-graphql.restart");
-  };
+  vscode.commands.registerCommand("fdc-graphql.start", async () => {
+    await generateYamlFile();
+    await client.start();
+    outputChannel.appendLine("Firebase GraphQL Language Server restarted");
+  });
 
+  // ** DISABLED FOR NOW WHILE WE TEST GENERATED YAML **
   // restart server whenever config file changes
-  const watcher = vscode.workspace.createFileSystemWatcher(
-    "**/.graphqlrc.*", // TODO: extend to schema files, and other config types
-    false,
-    false,
-    false
-  );
-  watcher.onDidChange(() => restartGraphqlLSP());
+  // const watcher = vscode.workspace.createFileSystemWatcher(
+  //   "**/.graphqlrc.*", // TODO: extend to schema files, and other config types
+  //   false,
+  //   false,
+  //   false,
+  // );
+  // watcher.onDidChange(() => restartGraphqlLSP());
 
   return client;
 }