+++ firebase-vscode/CHANGELOG.md
@@ -4,6 +4,8 @@
 
 - Update Extensions page Logo
 - Update README for Extensions page
+- Surface emulator issues as notifications
+- Generate .graphqlrc automatically
 - Emulator Bump 1.1.15
 
 ## 0.1.7